feat: add public allowedCharPattern property
